+++ apps/guiabolso-clone-server/src/api/syncTransactions.ts
@@ -1,4 +1,5 @@
-import * as SynchronizationService from "../services/synchronizationService";
+// import * as SynchronizationService from "../services/synchronizationService";
+import  { importTransactions } from '@gera-software/guiabolso-clone-core'
 import { AccountSummaryDTO } from "../types";
 import { Request, Response } from 'express'
 
@@ -10,7 +11,7 @@ export default async function(req: Request, res: Response) {
 
     let result: AccountSummaryDTO | null
     try {
-        result = await SynchronizationService.importTransactions(accountId, from)
+        result = await importTransactions(accountId, from)
     } catch(err) {
         console.error(err)
         return res.status(400).json(err)