+++ apps/guiabolso-clone-server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gera-software/guiabolso-clone-server
 
+## 0.3.0
+
+### Minor Changes
+
+- Using the core Syncrhonization service in the server
+
+### Patch Changes
+
+- Updated dependencies
+  - @gera-software/guiabolso-clone-core@0.1.0
+
 ## 0.2.1
 
 ### Patch Changes